requirements.txt
array-record==0.4.1
azure-core==1.32.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
Flask==3.1.0
Flask-Cors==5.0.0
requests==2.32.3
gunicorn